add cluster_ext
